requirements.txt
annotated-types==0.7.0
anyio>=4.7.0,<5.0.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
griffe==1.7.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
MarkupSafe==3.0.2
mcp==1.7.1
openai==1.77.0
openai-agents==0.0.14
PyPDF2==3.0.1
pypdfium2==4.30.1
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
sse-starlette==2.3.3
starlette==0.46.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
Werkzeug==3.1.3
fastapi>=0.109.0
# System dependency: poppler 25.06.0 required for pdf2image
# Install with: brew install poppler (macOS) or apt-get install poppler-utils (Ubuntu/Debian)
pdf2image==1.17.0
pillow==11.2.1
